+++ packages/angular-workspace/example-client-app/src/app/app.component.ts
@@ -5,7 +5,8 @@ import { Theme } from '@ni/nimble-angular';
     // eslint-disable-next-line @angular-eslint/component-selector
     selector: 'app-root',
     templateUrl: './app.component.html',
-    styleUrls: ['./app.component.scss']
+    styleUrls: ['./app.component.scss'],
+    standalone: false
 })
 export class AppComponent {
     public theme: Theme = Theme.light;