+++ packages/angular-workspace/nimble-angular/src/directives/tabs-toolbar/nimble-tabs-toolbar.directive.ts
@@ -8,7 +8,8 @@ export { tabsToolbarTag };
  * Directive to provide Angular integration for the tabs toolbar.
  */
 @Directive({
-    selector: 'nimble-tabs-toolbar'
+    selector: 'nimble-tabs-toolbar',
+    standalone: false
 })
 export class NimbleTabsToolbarDirective {
 }