+++ packages/angular-workspace/nimble-angular/src/directives/text-area/tests/nimble-text-area-control-value-accessor.directive.spec.ts
@@ -15,7 +15,8 @@ describe('Nimble text area control value accessor', () => {
     @Component({
         template: `
             <nimble-text-area #textArea [(ngModel)]="value" (ngModelChange)="onModelValueChange($event)" [disabled]="fieldDisabled"></nimble-text-area>
-         `
+         `,
+        standalone: false
     })
     class TestHostComponent {
         @ViewChild('textArea', { static: true }) public textArea: ElementRef<TextArea>;