+++ packages/angular-workspace/nimble-angular/src/directives/text-field/nimble-text-field-control-value-accessor.directive.ts
@@ -8,8 +8,7 @@ import { DefaultValueAccessor } from '../../thirdparty/directives/default_value_
  * Directive decorator based on DefaultValueAccessor decorator in thirdparty/directives/default_value_accessor
  */
 @Directive({
-    selector:
-        'nimble-text-field[formControlName],nimble-text-field[formControl],nimble-text-field[ngModel]',
+    selector: 'nimble-text-field[formControlName],nimble-text-field[formControl],nimble-text-field[ngModel]',
     // The following host metadata is duplicated from DefaultValueAccessor
     // eslint-disable-next-line @angular-eslint/no-host-metadata-property
     host: {
@@ -24,6 +23,7 @@ import { DefaultValueAccessor } from '../../thirdparty/directives/default_value_
         provide: NG_VALUE_ACCESSOR,
         useExisting: forwardRef(() => NimbleTextFieldControlValueAccessorDirective),
         multi: true
-    }]
+    }],
+    standalone: false
 })
 export class NimbleTextFieldControlValueAccessorDirective extends DefaultValueAccessor { }